fix(cli): fix bug when creating multiple bots in multiple Slack workspaces
Add a random hash to the end of created webtasks; each webtask will be
unique on the user's Webtask account. This provides a workaround for
when the same user adds a bot to different Slack workspaces. Previously,
the Webtask storage would be shared between Slack workspaces, causing
bugs with the ranking system.